+++ src/Main.java
@@ -1,7 +1,0 @@
-import stuman.StudentManager;
-
-public class Main {
-    public static void main(String[] args) {
-        StudentManager.initialize();
-    }
-}